opentelemetry_dev_requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
idna==3.11
importlib-metadata==8.7.1
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
protobuf==6.33.6
requests==2.32.5
typing-extensions==4.15.0
urllib3==1.26.20
zipp==3.23.0
